requirements.txt
requests==2.21.0
PyMySQL==0.9.3
grpcio==1.31.0
grpcio-tools==1.31.0
protobuf==3.12.4
pyyaml
tenacity==5.0.4